[DBTablesHosts] Integrate HostQueryOption and HostsQueryOption.
