+++ server/src/DBTablesMonitoring.cc
@@ -1447,77 +1447,6 @@ void ItemsQueryOption::setExcludeFlags(const ExcludeFlags &flg)
 }
 
 //
-// HostsQueryOption
-//
-static const HostResourceQueryOption::Synapse synapseHostsQueryOption(
-  tableProfileServerHostDef, IDX_HOST_SERVER_HOST_DEF_HOST_ID_IN_SERVER,
-  IDX_HOST_SERVER_HOST_DEF_SERVER_ID,
-  tableProfileServerHostDef, IDX_HOST_SERVER_HOST_DEF_HOST_ID_IN_SERVER,
-  true,
-  tableProfileHostgroupMember,
-  IDX_HOSTGROUP_MEMBER_SERVER_ID, IDX_HOSTGROUP_MEMBER_HOST_ID,
-  IDX_HOSTGROUP_MEMBER_GROUP_ID);
-
-struct HostsQueryOption::Impl {
-	HostStatus status;
-
-	Impl()
-	: status(HOST_STAT_ALL)
-	{
-	}
-};
-
-HostsQueryOption::HostsQueryOption(const UserIdType &userId)
-: HostResourceQueryOption(synapseHostsQueryOption, userId),
-  m_impl(new Impl())
-{
-}
-
-HostsQueryOption::HostsQueryOption(const HostsQueryOption &src)
-: HostResourceQueryOption(src),
-  m_impl(new Impl())
-{
-	*m_impl = *src.m_impl;
-}
-
-HostsQueryOption::HostsQueryOption(DataQueryContext *dataQueryContext)
-: HostResourceQueryOption(synapseHostsQueryOption, dataQueryContext),
-  m_impl(new Impl())
-{
-}
-
-HostsQueryOption::~HostsQueryOption(void)
-{
-}
-
-string HostsQueryOption::getCondition(void) const
-{
-	string condition = HostResourceQueryOption::getCondition();
-	if (m_impl->status == HOST_STAT_ALL)
-		return condition;
-	if (!condition.empty())
-		condition += " AND ";
-
-	const string columnName =
-	  tableProfileServerHostDef.getFullColumnName(
-	    IDX_HOST_SERVER_HOST_DEF_HOST_STATUS);
-	condition += StringUtils::sprintf("%s=%d",
-	                                  columnName.c_str(), m_impl->status);
-	return condition;
-}
-
-void HostsQueryOption::setStatus(const HostStatus &status)
-{
-	m_impl->status = status;
-}
-
-HostStatus HostsQueryOption::getStatus(void) const
-{
-	return m_impl->status;
-}
-
-
-//
 // HostgroupsQueryOption
 //
 static const HostResourceQueryOption::Synapse synapseHostgroupsQueryOption(
